+++ sdk/Javascript/src/Push/Push.WinJS.js
@@ -6,17 +6,17 @@
 /// <reference path="C:\Program Files (x86)\Microsoft SDKs\Windows\v8.0\ExtensionSDKs\Microsoft.WinJS.1.0\1.0\DesignTime\CommonConfiguration\Neutral\Microsoft.WinJS.1.0\js\ui.js" />
 /// <reference path="..\Generated\MobileServices.DevIntellisense.js" />
 
-var Validate = require('Validate');
-var LocalStorageManager = require('LocalStorageManager').LocalStorageManager;
-var RegistrationManager = require('RegistrationManager').RegistrationManager;
-var PushHttpClient = require('PushHttpClient').PushHttpClient;
+var _ = require('Extensions'),
+    Validate = require('Validate'),
+    Platform = require('Platform'),
+    RegistrationManager = require('RegistrationManager').RegistrationManager;
 
 function Push(mobileServicesClient, tileId) {
-    var localStorage = new LocalStorageManager(mobileServicesClient.applicationUrl, tileId);
-    this.registrationManager = new RegistrationManager(
-        new PushHttpClient(mobileServicesClient),
-        localStorage
-        );
+    tileId = tileId || '$Primary';
+    var packageName = Windows.ApplicationModel.Package.current.id.name,
+        name = _.format('{0}-PushContainer-{1}-{2}', packageName, mobileServicesClient.applicationUrl, tileId);
+
+    this._registrationManager = new RegistrationManager(mobileServicesClient, 'wns', name);
 }
 
 exports.Push = Push;
@@ -25,14 +25,19 @@ Push.prototype.registerNative = function (channelUri, tags) {
     /// <summary>
     /// Register for native notification
     /// </summary>
-    /// <param name="channelUri">The channelUri to register</param>
-    /// <param name="tags">Array containing the tags for this registeration</param>
-    /// <returns>Promise that will complete when the registration is completed</returns>
+    /// <param name="channelUri">
+    /// The channelUri to register
+    /// </param>
+    /// <param name="tags" mayBeNull="true">
+    /// Array containing the tags for this registeration
+    /// </param>
+    /// <returns>
+    /// A promise that will complete when the registration is complete.
+    /// </returns>
     
-    var registration = makeCoreRegistration(channelUri, tags);
-    return this.registrationManager.register(registration);
+    var registration = makeCoreRegistration(channelUri, tags);        
+    return this._registrationManager.upsertRegistration(registration);
 };
-
 
 Push.prototype.registerTemplate = function (channelUri, templateBody, templateName, headers, tags) {
     /// <summary>
@@ -43,8 +48,10 @@ Push.prototype.registerTemplate = function (channelUri, templateBody, templateNa
     /// <param name="templateName">The name of the template</param>
     /// <param name="headers">Object containing key/value pairs for the template to provide to WNS. X-WNS-Type is required. Example: { 'X-WNS-Type' : 'wns/toast' }</param>
     /// <param name="tags">Array containing the tags for this registeration</param>
-    /// <returns>Promise that will complete when the template registration is completed</returns>
-    
+    /// <returns>
+    /// A promise that will complete when the registration is complete.
+    /// </returns>
+
     var registration = makeCoreRegistration(channelUri, tags);
     
     if (templateBody) {
@@ -60,7 +67,7 @@ Push.prototype.registerTemplate = function (channelUri, templateBody, templateNa
         }
     }
 
-    return this.registrationManager.register(registration);
+    return this._registrationManager.upsertRegistration(registration);
 };
 
 Push.prototype.unregisterNative = function () {
@@ -69,33 +76,42 @@ Push.prototype.unregisterNative = function () {
     /// </summary>
     /// <returns>Promise that will complete when the unregister is completed</returns>
     
-    return this.unregisterTemplate(RegistrationManager.nativeRegistrationName);
+    return this.unregisterTemplate(RegistrationManager.NativeRegistrationName);
 };
 
 Push.prototype.unregisterTemplate = function (templateName) {
     /// <summary>
     /// Unregister for template notification
     /// </summary>
-    /// <param name="templateName">The name of the template</param>
+    /// <param name="templateName">
+    /// The name of the template
+    /// </param>
     /// <returns>Promise that will complete when the unregister is completed</returns>
     
-    Validate.notNullOrEmpty(templateName);
-    return this.registrationManager.unregister(templateName);
+    Validate.notNullOrEmpty(templateName, 'templateName');
+
+    return this._registrationManager.deleteRegistrationWithName(templateName);
 };
 
 Push.prototype.unregisterAll = function (channelUri) {
     /// <summary>
-    /// Unregister all notifications for a specfic channelUri
+    /// DEBUG-ONLY: Unregister all notifications for a specfic channelUri
     /// </summary>
     /// <param name="channelUri">The channelUri to unregister</param>
     /// <returns>Promise that will complete when the unregistration of all registrations at the channelUri is completed</returns>
     
-    Validate.notNullOrEmpty(channelUri);
-    return this.registrationManager.deleteRegistrationsForChannel(channelUri);
+    Validate.notNullOrEmpty(channelUri, 'channelUri');
+
+    return this._registrationManager.deleteAllRegistrations(channelUri);
 };
 
 Push.prototype.getSecondaryTile = function (tileId) {
-    return new Push(this.mobileServicesClient, tileId);
+    // TODO: move below to function
+    var packageName = Windows.ApplicationModel.Package.current.id.name,
+        name = _.format('{0}-PushContainer-{1}-{2}', packageName, mobileServicesClient.applicationUrl, tileId),
+        localStorage = new LocalStorageManager(name);
+
+    return new Push(name);
 };
 
 function makeCoreRegistration(channelUri, tags) {